move shellEnv to its own file, #16508
